Clarify the paging predicate chapter
There were some problems with that chapter. Namely,

- Explanation of the code sample was not clear. For example, ``values``
should refer to the method call, but it was used in plural form, probably
referring to the returned collection.
-  If the comparator is not set, keys or values of the collection should
be comparable, not the collection itself.
- The note about the transactional context is not related to the paragraphs
above. It makes sense to convert it into a note.

Credit goes to Andrey for spotting these problems on https://github.com/hazelcast/hazelcast-python-client/pull/232