Mention about the new changes to the connection retry config
As decided with the team, from now on, the client will not stop
trying to target cluster in the startup or after the disconnection
by default. For the failover client, by defualt, it will use 120000 ms
as the timeout value. If the user set this to something else, it will
be used as it is.

Also, mentioned about the client startup in the beginning of the
corresponding section because the configuration element also affects
that part.